Moved to using int positions in example.

+++ Examples/OpenGL/VertexShader.glsl
@@ -1,7 +1,7 @@
 #version 330 core
 
 // Input vertex data, different for all executions of this shader.
-layout(location = 0) in vec3 vertexPosition_modelspace;
+layout(location = 0) in uvec3 vertexPosition_modelspace;
 layout(location = 1) in vec2 vertexUV;
 
 // Output data ; will be interpolated for each fragment.
@@ -14,7 +14,6 @@ uniform mat4 projectionMatrix;
 
 void main()
 {
-
 	// Output position of the vertex, in clip space : MVP * position
 	gl_Position =  projectionMatrix * viewMatrix * modelMatrix * vec4(vertexPosition_modelspace,1);
